requirements.txt
# === Core Web Framework ===
fastapi==0.115.0
uvicorn==0.30.6

# === Google Cloud APIs ===
google-cloud-vision==3.7.4
google-cloud-translate==2.0.1
google-auth==2.35.0

# === Environment Configuration ===
python-dotenv==1.0.1

# === File Uploads ===
python-multipart==0.0.9

# === Utility / Text Parsing ===
regex==2024.9.11

# === Development / HTTP Tools (Optional) ===
requests==2.32.3